用户队列类中的属性类型升级为ObservableCollections中的类型
任务信息增加进度信息，并在任务无失败信息时显式
优化了部分代码逻辑与写法

+++ Services/Tasks/ITaskScheduler.cs
@@ -11,7 +11,7 @@ public interface ITaskScheduler : IDisposable
     /// <summary>
     /// 用户任务队列映射（用户ID -> UserTaskQueue）
     /// </summary>
-    ConcurrentDictionary<int, UserTaskQueue> UserQueues { get; }
+    ConcurrentDictionary<long, UserTaskQueue> UserQueues { get; }
 
     /// <summary>
     /// 最大并发协程数量
